Fixes UI version 1.4.2 Works with Marlin 2.1.2 with Input Shaping enabled.
UI version 1.4.2
is a fix for the UI. The Russian translation has been corrected. Fixed incorrect display of the language in the print confirmation menu. Added the display of coordinates in the MOVE menu. Fixed the FAN display in the print menu. The FAN speed value is shown as a percentage. Fixed the display of the Z coordinate in the print menu.